feat: 임시구현한 API에 web3j 연결 추가

Closes S12P21A604-186

See merge request s12-fintech-finance-sub1/S12P21A604!209

+++ rabbit-server/src/main/java/com/rabbit/loan/domain/dto/response/BorrowSummaryResponseDTO.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class BorrowSummaryResponseDTO {
 
-    private int totalOutgoingLa;
-    private int monthlyOutgoingLa;
+    private Long totalOutgoingLa;
+    private Long monthlyOutgoingLa;
     private String nextOutgoingDt;
 }
